third_party/WebKit/LayoutTests/bindings/webidl-operation-functions-should-have-no-prototype.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_false("prototype" in document.appendChild,
               "WebIDL operation functions should not have a prototype property.");
});
</script>
